+++ lib/lite_cable/server/client_socket/base.rb
@@ -70,13 +70,11 @@ module LiteCable
               @open_handlers.each(&:call)
               each_frame do |data|
                 @message_handlers.each do |h|
-                  begin
-                    h.call(data)
-                  rescue => e # rubocop: disable Style/RescueStandardError
-                    log(:error, "Socket receive failed: #{e}")
-                    @error_handlers.each { |eh| eh.call(e, data) }
-                    close if close_on_error
-                  end
+                  h.call(data)
+                rescue => e # rubocop: disable Style/RescueStandardError
+                  log(:error, "Socket receive failed: #{e}")
+                  @error_handlers.each { |eh| eh.call(e, data) }
+                  close if close_on_error
                 end
               end
             ensure