+++ lib/lite_cable/broadcast_adapters.rb
@@ -19,11 +19,11 @@ module LiteCable
           # We couldn't require the adapter itself.
           if e.path == path_to_adapter
             raise e.class, "Couldn't load the '#{adapter}' broadcast adapter for LiteCable",
-                  e.backtrace
+              e.backtrace
           # Bubbled up from the adapter require.
           else
             raise e.class, "Error loading the '#{adapter}' broadcast adapter for LiteCable",
-                  e.backtrace
+              e.backtrace
           end
         end
       end