+++ spec/spec_helper.rb
@@ -13,7 +13,7 @@ $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "lite_cable"
 require "lite_cable/server"
 
-Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
+Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].sort.each { |f| require f }
 
 RSpec.configure do |config|
   config.mock_with :rspec do |mocks|